envoy: move UpsertEnvoyEndpoints logic from xDSserver to CEC watcher
Currently, the xDSServer provides the possibility to update Envoy
endpoints based on loadbalancer backend information. This logic
is only used by the CiliumEnvoyConfig watcher to update the backend
services of a CEC accordingly.

With the introduction of a service backend sync callback, this logic
can be moved from the xDS server to the CEC watcher. This way we can
get rid of unwanted dependencies from the envoy module to the loadbalancer
module.
