+++ pkg/k8s/watchers/envoy_l7lb_backend_syncer_test.go
@@ -4,11 +4,16 @@
 package watchers
 
 import (
+	"net/netip"
 	"testing"
 
+	envoy_config_core "github.com/cilium/proxy/go/envoy/config/core/v3"
+	endpointv3 "github.com/cilium/proxy/go/envoy/config/endpoint/v3"
 	_ "github.com/cilium/proxy/go/envoy/config/listener/v3"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
+	"github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/loadbalancer"
 )
 
@@ -191,4 +196,49 @@ func Test_filterServiceBackends(t *testing.T) {
 			assert.Equal(t, (int)(backends["80"][0].Port), 8080)
 		})
 	})
+}
+
+func TestGetEndpointsForLBBackends(t *testing.T) {
+	testAddr, err := netip.ParseAddr("192.128.1.1")
+	require.NoError(t, err)
+
+	serviceName := loadbalancer.ServiceName{
+		Namespace: "test-ns",
+		Name:      "test-name",
+		Cluster:   "test-cluster",
+	}
+	backends := map[string][]*loadbalancer.Backend{
+		"12000": {
+			{
+				L3n4Addr: *loadbalancer.NewL3n4Addr(loadbalancer.TCP, types.AddrClusterFrom(testAddr, 0), 12000, 3),
+			},
+		},
+		"13000": {
+			{
+				L3n4Addr: *loadbalancer.NewL3n4Addr(loadbalancer.TCP, types.AddrClusterFrom(testAddr, 0), 13000, 3),
+			},
+		},
+		"*": {
+			{
+				L3n4Addr: *loadbalancer.NewL3n4Addr(loadbalancer.TCP, types.AddrClusterFrom(testAddr, 0), 15000, 3),
+			},
+		},
+	}
+
+	endpoints := getEndpointsForLBBackends(serviceName, backends)
+	assert.Len(t, endpoints, 4)
+
+	var allClusterNames []string
+	for _, ep := range endpoints {
+		allClusterNames = append(allClusterNames, ep.GetClusterName())
+
+		assert.Len(t, ep.GetEndpoints(), 1)
+		assert.Len(t, ep.GetEndpoints()[0].GetLbEndpoints(), 1)
+		assert.Equal(t, ep.GetEndpoints()[0].GetLbEndpoints()[0].GetHostIdentifier().(*endpointv3.LbEndpoint_Endpoint).Endpoint.Address.GetAddress().(*envoy_config_core.Address_SocketAddress).SocketAddress.Address, "192.128.1.1")
+	}
+
+	assert.Contains(t, allClusterNames, "test-cluster/test-ns/test-name:12000")
+	assert.Contains(t, allClusterNames, "test-cluster/test-ns/test-name:13000")
+	assert.Contains(t, allClusterNames, "test-cluster/test-ns/test-name:*")
+	assert.Contains(t, allClusterNames, "test-cluster/test-ns/test-name")
 }
